Implemented new ZNN.com simulation, according to SEAMS '13 paper.
Major changes were done to the meta-models, which might cause problems
to other existing simulations. I haven't tried them, so for now I will
only commit to the develop branch.

Also, the controller (and so also the remote API) have a new method
(EvoReq operation): doNothing(), so the simulation can be notified and
wake up when the "Do Nothing" strategy is used.

Other than this new strategy, new applicability conditions were also
created: FollowsSpecificStrategy, NotConcurrentWith and OrderInRange,
all of which are used in the ZNN.com simulation (see its model and
compare it with the paper to understand how they work).